refactor(knp-15): переименовал appCore в core, добавил правилf в линтер и добавил команду tsc
