+++ .gitignore
@@ -4,3 +4,4 @@
 /dist/*
 .env
 .env.*
+tsconfig.tsbuildinfo
