+++ .lintstagedrc
@@ -3,6 +3,7 @@
         "eslint . --ext js,jsx,ts,tsx,json --report-unused-disable-directives --max-warnings 0 --fix",
         "prettier --write \"**/*.+(js|jsx|ts|tsx|json|css|scss|less|sass)\""
     ],
+    "src/**/*.{ts,tsx}": ["npm run tsc"],
     "src/**/*.{css,scss,less,sass}": [
         "prettier --write \"**/*.+(js|jsx|ts|tsx|json|css|scss|less|sass)\""
     ]
