+++ package.json
@@ -6,11 +6,12 @@
         "start:dev": "cross-env NODE_ENV=development node server.js",
         "start:prod": "cross-env NODE_ENV=production node server.js",
         "build": "webpack --config webpack.prod.js",
-        "lint": "eslint . --ext js,jsx,ts,tsx,json --report-unused-disable-directives --max-warnings 0",
-        "lint:fix": "eslint . --ext js,jsx,ts,tsx,json --report-unused-disable-directives --max-warnings 0 --fix",
+        "lint": "npm run tsc && eslint . --ext js,jsx,ts,tsx,json --report-unused-disable-directives --max-warnings 0",
+        "lint:fix": "npm run tsc && eslint . --ext js,jsx,ts,tsx,json --report-unused-disable-directives --max-warnings 0 --fix",
         "prettier": "prettier --write \"**/*.+(js|jsx|ts|tsx|json|css|scss|less|sass)\"",
         "commit": "git-cz",
-        "precommit": "lint-staged -c .lintstagedrc"
+        "precommit": "lint-staged -c .lintstagedrc",
+        "tsc":"bash -c tsc --project tsconfig.json --noEmit --pretty"
     },
     "dependencies": {
         "express": "^4.18.2",
