+++ src/Providers/HistoryProvider.tsx
@@ -1,5 +1,5 @@
-import { AppComponent } from '@/appCore/src/AppComponent';
-import { Context } from '@/appCore/src/Context';
+import { AppComponent } from '@/core';
+import { Context } from '@/core/src/Context';
 import type { AppContext } from '@/types/Context.types';
 
 export interface HistoryContextProps {
