+++ src/core/app-dom/src/updateAttributes.ts
@@ -4,7 +4,7 @@ import {
     setAttributeWithCheck,
 } from './utils';
 
-import { APP_NODE_EVENTS } from '@/appCore/shared/AppNodeEvent.type';
+import { APP_NODE_EVENTS } from '@/core/shared/AppNodeEvent.type';
 import { includes, isDefined } from '@/utils';
 
 /**
