+++ src/core/app-dom/src/utils/getParentNodeFromElement.ts
@@ -1,4 +1,4 @@
-import type { AppNode } from '@/appCore/shared/AppNode.types';
+import type { AppNode } from '@/core/shared/AppNode.types';
 import { isPrimitive } from '@/utils';
 
 /**
