+++ src/core/app-dom/src/utils/getTextNode.ts
