+++ src/core/app-dom/src/utils/replaceChildWithCheck.ts
