+++ src/core/app-dom/src/utils/setAttributeWithCheck.ts
