+++ src/core/index.ts
@@ -1,0 +1,6 @@
+export { AppComponent } from './src/AppComponent';
+export type {
+    AppComponentType,
+    AppComponentConstructor,
+} from './src/AppComponent.types';
+export { Context } from './src/Context';
