+++ src/core/jsx-runtime/src/shared/JSXProps.type.ts
