+++ src/core/shared/AppNodeEvent.type.ts
