+++ src/core/shared/AppSymbols.ts
