+++ src/index.tsx
@@ -1,4 +1,4 @@
 import { App } from '@/App';
-import { createRoot } from '@/appCore/app-dom';
+import { createRoot } from '@/core/app-dom';
 
 createRoot(document.getElementById('root'), <App />);
