+++ src/pages/IndexPage/IndexPage.tsx
@@ -1,4 +1,4 @@
-import { AppComponent } from '@/appCore/src/AppComponent';
+import { AppComponent } from '@/core';
 import { FilmsContainer } from '@/components/FilmsContainer';
 import { LayoutWithHeader } from '@/layouts/LayoutWithHeader';
