+++ src/utils/spreadToSingleArray.ts
@@ -1,5 +1,3 @@
-type Props<T> = T | (Props<T> | Props<T>[])[];
-
 /**
  * Разворачивает массив массивов в один массив
  * @param value Массив, который может состоять из
@@ -7,13 +5,15 @@ type Props<T> = T | (Props<T> | Props<T>[])[];
  * @returns Возвращает один массив, состоящий только из элементов не-массивов
  * @example `spreadToSingleArray([1, [[2, 3], 4], 5]) === [1, 2, 3, 4, 5]`
  */
-export const spreadToSingleArray = <T>(value: Props<T>) => {
+export const spreadToSingleArray = <T>(value: T | Iterable<T>) => {
     let array: T[] = [];
 
-    if (Array.isArray(value)) {
-        value.forEach((child) => {
-            array = [...array, ...spreadToSingleArray(child)];
-        });
+    if (value && typeof value === 'object' && Symbol.iterator in value) {
+        if (Array.isArray(value)) {
+            value.forEach((child: T) => {
+                array = [...array, ...spreadToSingleArray(child)];
+            });
+        }
     } else {
         array.push(value);
     }
